requirements.txt
# requirements.txt

# --- Core Dependencies ---
requests
pandas
pyarrow


# --- Google Cloud Dependencies ---
google-cloud-storage
google-cloud-bigquery
gcsfs